Decouple `CallbackOnce` from `CallbackStream`
